+++ packages/lesswrong/components/common/ContentItemTruncated.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Components, registerComponent } from '../../lib/vulcan-lib';
-import { truncate } from '../../lib/editor/ellipsize';
+import { truncateWithGrace } from '../../lib/editor/ellipsize';
 import classNames from 'classnames';
 
 const styles = (theme: ThemeType): JssStyles => ({
@@ -49,28 +49,6 @@ const ContentItemTruncated = ({classes, maxLengthWords, graceWords=20, expanded=
     {wasTruncated && getTruncatedSuffix && getTruncatedSuffix({wordsLeft})}
     {!wasTruncated && nonTruncatedSuffix}
   </>
-}
-
-export const truncateWithGrace = (html: string, maxLengthWords: number, graceWords: number, rawWordCount: number, suffix?: string): {
-  truncatedHtml: string,
-  wasTruncated: boolean,
-  wordsLeft: number,
-} => {
-  const truncatedHtml = truncate(html, maxLengthWords, "words", suffix);
-  const wordsLeft = (truncatedHtml===html) ? 0 : rawWordCount-maxLengthWords;
-  
-  if (truncatedHtml === html || wordsLeft<graceWords) {
-    return {
-      truncatedHtml: html,
-      wasTruncated: false,
-      wordsLeft: 0,
-    };
-  }
-  
-  return {
-    truncatedHtml, wordsLeft,
-    wasTruncated: true,
-  };
 }
 
 const ContentItemTruncatedComponent = registerComponent('ContentItemTruncated', ContentItemTruncated, {styles});
